Create Scala 2 veneer for Instantiate.scala
